+++ Assets/OkapiKit/Scripts/Editor/OkapiTargetInventoryEditor.cs
@@ -1,8 +1,0 @@
-using OkapiKit.Editor;
-using OkapiKit;
-using UnityEditor;
-
-[CustomPropertyDrawer(typeof(TargetInventory))]
-public class TargetInventoryEditor : OkapiTargetEditor<Inventory>
-{
-}